third_party/WebKit/LayoutTests/media/track/track-remove-insert-ready-state-expected.txt

Tests that re-adding a media element to the document, having a child track that failed loading doesn't block video from playing

** Track should have error as ready state **
EXPECTED (track.readyState == HTMLTrackElement.ERROR == 'true') OK

** Remove the video element from body **
** Reset the video src attribute to re-trigger resource selection for tracks **
** Append the video element back to the body **

** The video should start playing **
EXPECTED (track.readyState == HTMLTrackElement.ERROR == 'true') OK

** Succesfully played the video **
